Progress is being made on TC/TM transactions. I'm seeing some weird stuff on the incoming CAN messages. So far I've been able to receive the TC_PKT_RDY message and I can send back the OKAY_START_TC_TRANSFER message.

+++ Subsytem_Code/Code/global_var.h
@@ -274,7 +274,7 @@ uint8_t LOW_POWER_MODE;
 uint8_t PAUSE;
 
 /* Global variables used for operational timeouts */
-uint8_t ssm_ok_go_timeout;
+uint32_t ssm_ok_go_timeout;
 uint8_t ssm_consec_trans_timeout;
 
 /* Signal for FDIR Error Handling (SSM Loops on this) */
